task_module/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
django-cors-headers
djangorestframework-simplejwt==5.2.2
redis==4.5.5
django-redis==5.2.0
drf-spectacular==0.27.1  # Добавьте эту строку
